Ok, so we were having problems when submitting a form with a date field, but the date field wasn't in the post data.  FormEncode was expecting a string instance and was throwing an error b/c NotGiven is not a string.
So, I just made NotGivenBase a subclass of unicode.  That was ok, but caused some other problems where we were testing with for strings using basestring, which now NotGiven was included in.  That necessitated the is_str() function.

Also, now that NotGiven == '', FormEncode processors will return None every time they see it.  However, we actually want the NotGiven value, so we had to do a workaround there.